JZL1CSS-19 - refactor the chooseCarType menu
